+++ crates/aiken-project/src/error.rs
@@ -6,15 +6,20 @@ use aiken_lang::{
     test_framework::{BenchmarkResult, PropertyTestResult, TestResult, UnitTestResult},
     tipo,
 };
+use hex::FromHexError;
+use indoc::formatdoc;
 use miette::{
     Diagnostic, EyreContext, LabeledSpan, MietteHandler, MietteHandlerOpts, NamedSource, RgbColors,
     SourceCode,
 };
+use ordinal::Ordinal;
 use owo_colors::{
     OwoColorize,
     Stream::{Stderr, Stdout},
 };
+use pallas_addresses::ScriptHash;
 use std::{
+    collections::BTreeSet,
     fmt::{self, Debug, Display},
     io,
     path::{Path, PathBuf},
@@ -149,6 +154,120 @@ pub enum Error {
 
     #[error("I located conditional modules under 'env', but no default one!")]
     NoDefaultEnvironment,
+
+    #[error(
+        "I couldn't find any script matching {} in your blueprint (plutus.json).",
+        script_hash.to_string().if_supports_color(Stdout, |s| s.yellow()),
+    )]
+    ScriptOverrideNotFound {
+        script_hash: ScriptHash,
+        known_scripts: BTreeSet<ScriptHash>,
+    },
+
+    #[error(
+        "I couldn't parse the {} script override argument.",
+        Ordinal(*index as u32 + 1).to_string().if_supports_color(Stdout, |s| s.bold()),
+    )]
+    ScriptOverrideArgumentParseError {
+        index: usize,
+        #[source]
+        error: ScriptOverrideArgumentError,
+    },
+}
+
+#[derive(thiserror::Error, Debug, Diagnostic)]
+pub enum ScriptOverrideArgumentError {
+    #[error(
+        "I couldn't find the left side of the mapping. Are you sure you provided two colon-separated validator qualifiers?\n"
+    )]
+    #[diagnostic(help(
+        "{}",
+        formatdoc!{
+            r#"I am expecting a mapping {from} a validator of the transaction, {to} a replacement in the blueprint as such:
+
+                --script-override "122171C7C82348C420A47AFD8F{colon}7730C913CEEC22524B98E1604A"
+                                   {underline}
+                                    {missing}
+            "#,
+            from = "from"
+                .if_supports_color(Stdout, |s| s.bold()),
+            to = "to"
+                .if_supports_color(Stdout, |s| s.bold()),
+            colon = ":"
+                .if_supports_color(Stderr, |s| s.cyan())
+                .if_supports_color(Stderr, |s| s.bold()),
+            underline = "^^^^^^^^^^^^^^^^^^^^^^^^^^"
+                .if_supports_color(Stdout, |s| s.yellow())
+                .if_supports_color(Stdout, |s| s.bold()),
+            missing = "missing left side (from)"
+                .if_supports_color(Stdout, |s| s.yellow())
+        },
+    ))]
+    MissingFrom,
+
+    #[error(
+        "I couldn't find the right side of the mapping. Are you sure you provided two colon-separated validator qualifiers?\n"
+    )]
+    #[diagnostic(help(
+        "{}",
+        formatdoc!{
+            r#"I am expecting a mapping {from} a validator of the transaction, {to} a replacement in the blueprint as such:
+
+                --script-override "122171C7C82348C420A47AFD8F{colon}7730C913CEEC22524B98E1604A"
+                                                              {underline}
+                                                               {missing}
+            "#,
+            from = "from"
+                .if_supports_color(Stdout, |s| s.bold()),
+            to = "to"
+                .if_supports_color(Stdout, |s| s.bold()),
+            colon = ":"
+                .if_supports_color(Stderr, |s| s.cyan())
+                .if_supports_color(Stderr, |s| s.bold()),
+            underline = "^^^^^^^^^^^^^^^^^^^^^^^^^^"
+                .if_supports_color(Stdout, |s| s.yellow())
+                .if_supports_color(Stdout, |s| s.bold()),
+            missing = "missing right side (to)"
+                .if_supports_color(Stdout, |s| s.yellow())
+        },
+    ))]
+    MissingTo,
+
+    #[error(
+        "The origin qualifier ({}) isn't a valid hex-encoded hash digest.\n",
+        "from".if_supports_color(Stdout, |s| s.bold()),
+    )]
+    #[diagnostic(help("{0}"))]
+    InvalidFromHash(FromHexError),
+
+    #[error(
+        "The destination qualifier ({}) isn't a valid hex-encoded hash digest.\n",
+        "to".if_supports_color(Stdout, |s| s.bold()),
+    )]
+    #[diagnostic(help("{0}"))]
+    InvalidToHash(FromHexError),
+
+    #[error(
+        "The origin qualifier ({}) isn't a valid script hash digest.\n",
+        "from".if_supports_color(Stdout, |s| s.bold()),
+    )]
+    #[diagnostic(help(
+        "The size is incorrect. I expected to find {} bytes but found {}.",
+        "28".if_supports_color(Stdout, |s| s.green()),
+        .0.to_string().if_supports_color(Stdout, |s| s.red()),
+    ))]
+    InvalidFromSize(usize),
+
+    #[error(
+        "The destination qualifier ({}) isn't a valid script hash digest.\n",
+        "to".if_supports_color(Stdout, |s| s.bold()),
+    )]
+    #[diagnostic(help(
+        "The size is incorrect. I expected to find {} bytes but found {}.",
+        "28".if_supports_color(Stdout, |s| s.green()),
+        .0.to_string().if_supports_color(Stdout, |s| s.red()),
+    ))]
+    InvalidToSize(usize),
 }
 
 impl Error {
@@ -244,7 +363,9 @@ impl ExtraData for Error {
             | Error::Module { .. }
             | Error::NoDefaultEnvironment
             | Error::ModuleNotFound { .. }
-            | Error::ExportNotFound { .. } => None,
+            | Error::ExportNotFound { .. }
+            | Error::ScriptOverrideNotFound { .. }
+            | Error::ScriptOverrideArgumentParseError { .. } => None,
             Error::Type { error, .. } => error.extra_data(),
         }
     }
@@ -273,7 +394,9 @@ impl GetSource for Error {
             | Error::ModuleNotFound { .. }
             | Error::ExportNotFound { .. }
             | Error::NoDefaultEnvironment
-            | Error::Module { .. } => None,
+            | Error::Module { .. }
+            | Error::ScriptOverrideNotFound { .. }
+            | Error::ScriptOverrideArgumentParseError { .. } => None,
             Error::DuplicateModule { second: path, .. }
             | Error::MissingManifest { path }
             | Error::TomlLoading { path, .. }
@@ -303,7 +426,9 @@ impl GetSource for Error {
             | Error::NoDefaultEnvironment
             | Error::ModuleNotFound { .. }
             | Error::ExportNotFound { .. }
-            | Error::Module { .. } => None,
+            | Error::Module { .. }
+            | Error::ScriptOverrideNotFound { .. }
+            | Error::ScriptOverrideArgumentParseError { .. } => None,
             Error::TomlLoading { src, .. } | Error::Parse { src, .. } | Error::Type { src, .. } => {
                 Some(src.to_string())
             }
@@ -329,7 +454,6 @@ impl Diagnostic for Error {
 
         match self {
             Error::DuplicateModule { .. } => Some(boxed(Box::new("aiken::module::duplicate"))),
-            Error::FileIo { .. } => None,
             Error::Blueprint(e) => e.code().map(boxed),
             Error::ImportCycle { .. } => Some(boxed(Box::new("aiken::module::cyclical"))),
             Error::Parse { .. } => Some(boxed(Box::new("aiken::parser"))),
@@ -337,25 +461,28 @@ impl Diagnostic for Error {
                 "aiken::check{}",
                 error.code().map(|s| format!("::{s}")).unwrap_or_default()
             )))),
-            Error::StandardIo(_) => None,
-            Error::MissingManifest { .. } => None,
             Error::TomlLoading { .. } => Some(boxed(Box::new("aiken::loading::toml"))),
-            Error::Format { .. } => None,
             Error::TestFailure { path, .. } => Some(boxed(Box::new(path.to_str().unwrap_or("")))),
             Error::Http(_) => Some(Box::new("aiken::packages::download")),
-            Error::ZipExtract(_) => None,
-            Error::JoinError(_) => None,
             Error::UnknownPackageVersion { .. } => {
                 Some(boxed(Box::new("aiken::packages::resolve")))
             }
             Error::UnableToResolvePackage { .. } => {
                 Some(boxed(Box::new("aiken::package::download")))
             }
-            Error::Json { .. } => None,
-            Error::MalformedStakeAddress { .. } => None,
-            Error::ExportNotFound { .. } => None,
-            Error::ModuleNotFound { .. } => None,
-            Error::NoDefaultEnvironment => None,
+            Error::StandardIo(_)
+            | Error::MissingManifest { .. }
+            | Error::ZipExtract(_)
+            | Error::JoinError(_)
+            | Error::FileIo { .. }
+            | Error::Format { .. }
+            | Error::Json { .. }
+            | Error::MalformedStakeAddress { .. }
+            | Error::ExportNotFound { .. }
+            | Error::ModuleNotFound { .. }
+            | Error::NoDefaultEnvironment
+            | Error::ScriptOverrideNotFound { .. }
+            | Error::ScriptOverrideArgumentParseError { .. } => None,
             Error::Module(e) => e.code().map(boxed),
         }
     }
@@ -375,7 +502,6 @@ impl Diagnostic for Error {
             ))),
             Error::Parse { error, .. } => error.help(),
             Error::Type { error, .. } => error.help(),
-            Error::StandardIo(_) => None,
             Error::MissingManifest { .. } => Some(Box::new(
                 "Try running `aiken new <REPOSITORY/PROJECT>` to initialise a project with an example manifest.",
             )),
@@ -383,12 +509,7 @@ impl Diagnostic for Error {
                 "Environment module names are free, but there must be at least one named 'default.ak'.",
             )),
             Error::TomlLoading { help, .. } => Some(Box::new(help)),
-            Error::Format { .. } => None,
-            Error::TestFailure { .. } => None,
-            Error::Http(_) => None,
-            Error::ZipExtract(_) => None,
-            Error::JoinError(_) => None,
-            Error::ExportNotFound { .. } => None,
+
             Error::ModuleNotFound { known_modules, .. } => Some(Box::new(format!(
                 "I know about the following modules:\n{}",
                 known_modules
@@ -411,21 +532,31 @@ impl Diagnostic for Error {
                     None => String::new(),
                 }
             ))),
+            Error::ScriptOverrideNotFound { known_scripts, .. } => Some(Box::new(format!(
+                "These are all the scripts found in your blueprint:\n\n{}",
+                known_scripts
+                    .iter()
+                    .map(|s| format!("✓ {s}"))
+                    .collect::<Vec<_>>()
+                    .join("\n"),
+            ))),
+            Error::ScriptOverrideArgumentParseError { error, .. } => error.help(),
             Error::Module(e) => e.help(),
+            Error::StandardIo(_)
+            | Error::Format { .. }
+            | Error::TestFailure { .. }
+            | Error::Http(_)
+            | Error::ZipExtract(_)
+            | Error::JoinError(_)
+            | Error::ExportNotFound { .. } => None,
         }
     }
 
     fn labels(&self) -> Option<Box<dyn Iterator<Item = LabeledSpan> + '_>> {
         match self {
-            Error::DuplicateModule { .. } => None,
-            Error::FileIo { .. } => None,
-            Error::ImportCycle { .. } => None,
-            Error::ExportNotFound { .. } => None,
             Error::Blueprint(e) => e.labels(),
             Error::Parse { error, .. } => error.labels(),
-            Error::MissingManifest { .. } => None,
             Error::Type { error, .. } => error.labels(),
-            Error::StandardIo(_) => None,
             Error::TomlLoading { location, .. } => {
                 if let Some(location) = location {
                     Some(Box::new(
@@ -435,98 +566,114 @@ impl Diagnostic for Error {
                     None
                 }
             }
-            Error::Format { .. } => None,
-            Error::TestFailure { .. } => None,
-            Error::Http(_) => None,
-            Error::ZipExtract(_) => None,
-            Error::JoinError(_) => None,
-            Error::UnknownPackageVersion { .. } => None,
-            Error::UnableToResolvePackage { .. } => None,
-            Error::Json { .. } => None,
-            Error::MalformedStakeAddress { .. } => None,
-            Error::NoDefaultEnvironment => None,
-            Error::ModuleNotFound { .. } => None,
+            Error::DuplicateModule { .. }
+            | Error::FileIo { .. }
+            | Error::ImportCycle { .. }
+            | Error::ExportNotFound { .. }
+            | Error::StandardIo(_)
+            | Error::MissingManifest { .. }
+            | Error::Format { .. }
+            | Error::TestFailure { .. }
+            | Error::Http(_)
+            | Error::ZipExtract(_)
+            | Error::JoinError(_)
+            | Error::UnknownPackageVersion { .. }
+            | Error::UnableToResolvePackage { .. }
+            | Error::Json { .. }
+            | Error::MalformedStakeAddress { .. }
+            | Error::NoDefaultEnvironment
+            | Error::ModuleNotFound { .. }
+            | Error::ScriptOverrideNotFound { .. }
+            | Error::ScriptOverrideArgumentParseError { .. } => None,
+
             Error::Module(e) => e.labels(),
         }
     }
 
     fn source_code(&self) -> Option<&dyn SourceCode> {
         match self {
-            Error::DuplicateModule { .. } => None,
-            Error::FileIo { .. } => None,
-            Error::ImportCycle { .. } => None,
-            Error::ModuleNotFound { .. } => None,
-            Error::ExportNotFound { .. } => None,
             Error::Blueprint(e) => e.source_code(),
-            Error::NoDefaultEnvironment => None,
             Error::Parse { named, .. } => Some(named.as_ref()),
             Error::Type { named, .. } => Some(named),
-            Error::StandardIo(_) => None,
-            Error::MissingManifest { .. } => None,
             Error::TomlLoading { named, .. } => Some(named.as_ref()),
-            Error::Format { .. } => None,
-            Error::TestFailure { .. } => None,
-            Error::Http(_) => None,
-            Error::ZipExtract(_) => None,
-            Error::JoinError(_) => None,
-            Error::UnknownPackageVersion { .. } => None,
-            Error::UnableToResolvePackage { .. } => None,
-            Error::Json { .. } => None,
-            Error::MalformedStakeAddress { .. } => None,
+            Error::DuplicateModule { .. }
+            | Error::FileIo { .. }
+            | Error::ImportCycle { .. }
+            | Error::ModuleNotFound { .. }
+            | Error::ExportNotFound { .. }
+            | Error::NoDefaultEnvironment
+            | Error::StandardIo(_)
+            | Error::MissingManifest { .. }
+            | Error::Format { .. }
+            | Error::TestFailure { .. }
+            | Error::Http(_)
+            | Error::ZipExtract(_)
+            | Error::JoinError(_)
+            | Error::UnknownPackageVersion { .. }
+            | Error::UnableToResolvePackage { .. }
+            | Error::Json { .. }
+            | Error::MalformedStakeAddress { .. }
+            | Error::ScriptOverrideNotFound { .. }
+            | Error::ScriptOverrideArgumentParseError { .. } => None,
             Error::Module(e) => e.source_code(),
         }
     }
 
     fn url<'a>(&'a self) -> Option<Box<dyn Display + 'a>> {
         match self {
-            Error::DuplicateModule { .. } => None,
-            Error::FileIo { .. } => None,
-            Error::ImportCycle { .. } => None,
-            Error::ModuleNotFound { .. } => None,
-            Error::ExportNotFound { .. } => None,
             Error::Blueprint(e) => e.url(),
-            Error::Parse { .. } => None,
             Error::Type { error, .. } => error.url(),
-            Error::StandardIo(_) => None,
-            Error::MissingManifest { .. } => None,
-            Error::TomlLoading { .. } => None,
-            Error::Format { .. } => None,
-            Error::TestFailure { .. } => None,
-            Error::Http { .. } => None,
-            Error::ZipExtract { .. } => None,
-            Error::JoinError { .. } => None,
-            Error::UnknownPackageVersion { .. } => None,
-            Error::UnableToResolvePackage { .. } => None,
-            Error::Json { .. } => None,
-            Error::MalformedStakeAddress { .. } => None,
-            Error::NoDefaultEnvironment => None,
+            Error::DuplicateModule { .. }
+            | Error::FileIo { .. }
+            | Error::ImportCycle { .. }
+            | Error::ModuleNotFound { .. }
+            | Error::ExportNotFound { .. }
+            | Error::Parse { .. }
+            | Error::StandardIo(_)
+            | Error::MissingManifest { .. }
+            | Error::TomlLoading { .. }
+            | Error::Format { .. }
+            | Error::TestFailure { .. }
+            | Error::Http { .. }
+            | Error::ZipExtract { .. }
+            | Error::JoinError { .. }
+            | Error::UnknownPackageVersion { .. }
+            | Error::UnableToResolvePackage { .. }
+            | Error::Json { .. }
+            | Error::MalformedStakeAddress { .. }
+            | Error::NoDefaultEnvironment
+            | Error::ScriptOverrideNotFound { .. }
+            | Error::ScriptOverrideArgumentParseError { .. } => None,
+
             Error::Module(e) => e.url(),
         }
     }
 
     fn related<'a>(&'a self) -> Option<Box<dyn Iterator<Item = &'a dyn Diagnostic> + 'a>> {
         match self {
-            Error::DuplicateModule { .. } => None,
-            Error::FileIo { .. } => None,
-            Error::ModuleNotFound { .. } => None,
-            Error::ExportNotFound { .. } => None,
             Error::Blueprint(e) => e.related(),
-            Error::ImportCycle { .. } => None,
-            Error::Parse { .. } => None,
             Error::Type { error, .. } => error.related(),
-            Error::StandardIo(_) => None,
-            Error::NoDefaultEnvironment => None,
-            Error::MissingManifest { .. } => None,
-            Error::TomlLoading { .. } => None,
-            Error::Format { .. } => None,
-            Error::TestFailure { .. } => None,
-            Error::Http { .. } => None,
-            Error::ZipExtract { .. } => None,
-            Error::JoinError { .. } => None,
-            Error::UnknownPackageVersion { .. } => None,
-            Error::UnableToResolvePackage { .. } => None,
-            Error::Json { .. } => None,
-            Error::MalformedStakeAddress { .. } => None,
+            Error::DuplicateModule { .. }
+            | Error::FileIo { .. }
+            | Error::ModuleNotFound { .. }
+            | Error::ExportNotFound { .. }
+            | Error::ImportCycle { .. }
+            | Error::Parse { .. }
+            | Error::StandardIo(_)
+            | Error::NoDefaultEnvironment
+            | Error::MissingManifest { .. }
+            | Error::TomlLoading { .. }
+            | Error::Format { .. }
+            | Error::TestFailure { .. }
+            | Error::Http { .. }
+            | Error::ZipExtract { .. }
+            | Error::JoinError { .. }
+            | Error::UnknownPackageVersion { .. }
+            | Error::UnableToResolvePackage { .. }
+            | Error::Json { .. }
+            | Error::MalformedStakeAddress { .. }
+            | Error::ScriptOverrideNotFound { .. }
+            | Error::ScriptOverrideArgumentParseError { .. } => None,
             Error::Module(e) => e.related(),
         }
     }