+++ crates/aiken/src/cmd/tx/mod.rs
@@ -1,6 +1,7 @@
 pub mod simulate;
 
 use clap::Subcommand;
+use std::process;
 
 /// Commands for working with transactions
 #[derive(Subcommand)]
@@ -13,4 +14,5 @@ pub fn exec(cmd: Cmd) -> miette::Result<()> {
     match cmd {
         Cmd::Simulate(args) => simulate::exec(args),
     }
+    .map_err(|_| process::exit(1))
 }