+++ crates/aiken/src/cmd/tx/simulate.rs
@@ -1,16 +1,24 @@
+use aiken_project::{
+    error::{Error, ScriptOverrideArgumentError},
+    telemetry::EventTarget,
+    watch::with_project,
+};
+use hex::FromHexError;
 use miette::IntoDiagnostic;
 use owo_colors::{OwoColorize, Stream::Stderr};
+use pallas_addresses::ScriptHash;
 use pallas_primitives::{
     Fragment,
     conway::{Redeemer, TransactionInput, TransactionOutput},
 };
 use pallas_traverse::{Era, MultiEraTx};
-use std::{fmt, fs, path::PathBuf, process};
+use serde_json::json;
+use std::{collections::HashMap, fmt, fs, path::PathBuf, process};
 use uplc::{
-    machine::cost_model::ExBudget,
+    machine::Trace,
     tx::{
         self, redeemer_tag_to_string,
-        script_context::{ResolvedInput, SlotConfig},
+        script_context::{PlutusScript, ResolvedInput, SlotConfig},
     },
 };
 
@@ -45,6 +53,15 @@ pub struct Args {
     /// Slot number at the start of the shelley hardfork
     #[clap(long, default_value_t = 4492800, value_name = "SLOT")]
     zero_slot: u64,
+
+    /// An Aiken blueprint JSON file containing the overriding scripts, if applicable
+    #[clap(long, value_name = "FILEPATH")]
+    blueprint: Option<PathBuf>,
+
+    /// A mapping (colon-separated) from a script hash (in the transaction) to override by another script found in the blueprint.
+    /// For example:`d27cee75:197c9353`
+    #[clap(long("script-override"), value_name = "FROM:TO", num_args(0..), verbatim_doc_comment)]
+    script_overrides: Vec<String>,
 }
 
 pub fn exec(
@@ -56,6 +73,8 @@ pub fn exec(
         slot_length,
         zero_time,
         zero_slot,
+        blueprint,
+        script_overrides,
     }: Args,
 ) -> miette::Result<()> {
     eprintln!(
@@ -85,9 +104,78 @@ pub fn exec(
 
     let tx = MultiEraTx::decode_for_era(Era::Conway, &tx_bytes).into_diagnostic()?;
 
+    let mut overrides: HashMap<ScriptHash, PlutusScript> = HashMap::new();
+
+    if !script_overrides.is_empty() {
+        with_project(None, false, true, false, |p| {
+            eprintln!(
+                "{:>13} scripts",
+                "Overriding"
+                    .if_supports_color(Stderr, |s| s.purple())
+                    .if_supports_color(Stderr, |s| s.bold()),
+            );
+
+            let blueprint_path = p.blueprint_path(blueprint.as_deref());
+            let blueprint = p.blueprint(&blueprint_path)?;
+            let blueprint_validators: HashMap<ScriptHash, PlutusScript> = blueprint.into();
+
+            script_overrides
+                .iter()
+                .enumerate()
+                .try_for_each::<_, Result<_, aiken_project::error::Error>>(
+                    |(index, script_override)| {
+                        let mut parts = script_override.split(":");
+
+                        let from =
+                            get_override_part(&mut parts, ScriptOverrideArgumentError::MissingFrom)
+                                .and_then(|part| {
+                                    decode_script_hash(
+                                        part,
+                                        ScriptOverrideArgumentError::InvalidFromHash,
+                                        ScriptOverrideArgumentError::InvalidFromSize,
+                                    )
+                                })
+                                .map_err(|error| Error::ScriptOverrideArgumentParseError {
+                                    index,
+                                    error,
+                                })?;
+
+                        let to =
+                            get_override_part(&mut parts, ScriptOverrideArgumentError::MissingTo)
+                                .and_then(|part| {
+                                    decode_script_hash(
+                                        part,
+                                        ScriptOverrideArgumentError::InvalidToHash,
+                                        ScriptOverrideArgumentError::InvalidToSize,
+                                    )
+                                })
+                                .map_err(|error| Error::ScriptOverrideArgumentParseError {
+                                    index,
+                                    error,
+                                })?;
+
+                        overrides.insert(
+                            from,
+                            blueprint_validators
+                                .get(&to)
+                                .ok_or_else(|| Error::ScriptOverrideNotFound {
+                                    script_hash: to,
+                                    known_scripts: blueprint_validators.keys().cloned().collect(),
+                                })?
+                                .clone(),
+                        );
+
+                        Ok(())
+                    },
+                )?;
+
+            Ok(())
+        })?;
+    }
+
     eprintln!(
-        "{} {}",
-        "   Simulating"
+        "{:>13} {}",
+        "Simulating"
             .if_supports_color(Stderr, |s| s.purple())
             .if_supports_color(Stderr, |s| s.bold()),
         tx.hash()
@@ -114,8 +202,8 @@ pub fn exec(
 
         let with_redeemer = |redeemer: &Redeemer| {
             eprintln!(
-                "{} {}[{}]",
-                "   Evaluating"
+                "{:>13} {}[{}]",
+                "Evaluating"
                     .if_supports_color(Stderr, |s| s.purple())
                     .if_supports_color(Stderr, |s| s.bold()),
                 redeemer_tag_to_string(&redeemer.tag),
@@ -123,39 +211,93 @@ pub fn exec(
             )
         };
 
-        let result = tx::eval_phase_two(
+        let result = tx::eval_phase_two_with_override(
             tx_conway,
             &resolved_inputs,
             None,
             None,
             &slot_config,
+            overrides,
             true,
             with_redeemer,
         );
 
         match result {
             Ok(redeemers) => {
-                // this should allow N scripts to be
-                let total_budget_used: Vec<ExBudget> = redeemers
-                    .iter()
-                    .map(|(curr, _)| ExBudget {
-                        mem: curr.ex_units.mem as i64,
-                        cpu: curr.ex_units.steps as i64,
-                    })
-                    .collect();
+                let is_terminal = matches!(EventTarget::default(), EventTarget::Terminal(..));
 
-                eprintln!("\n");
-                println!(
-                    "{}",
-                    serde_json::to_string_pretty(&total_budget_used)
-                        .map_err(|_| fmt::Error)
-                        .into_diagnostic()?
-                );
+                if is_terminal {
+                    redeemers.iter().for_each(|(redeemer, eval)| {
+                        eprintln!(
+                            "{} {}={}\n{:>6}{:>11}={}",
+                            format!(
+                                "{:>13}",
+                                format!(
+                                    "{}[{}]",
+                                    redeemer_tag_to_string(&redeemer.tag),
+                                    redeemer.index
+                                )
+                            )
+                            .if_supports_color(Stderr, |s| s.purple())
+                            .if_supports_color(Stderr, |s| s.bold()),
+                            "mem".if_supports_color(Stderr, |s| s.bold()),
+                            redeemer.ex_units.mem,
+                            "│"
+                                .if_supports_color(Stderr, |s| s.purple())
+                                .if_supports_color(Stderr, |s| s.bold()),
+                            "cpu".if_supports_color(Stderr, |s| s.bold()),
+                            redeemer.ex_units.steps,
+                        );
+
+                        let traces = eval.traces();
+
+                        eprintln!(
+                            "{:>13} {}",
+                            "└ Traces"
+                                .if_supports_color(Stderr, |s| s.purple())
+                                .if_supports_color(Stderr, |s| s.bold()),
+                            if traces.is_empty() {
+                                "ø".to_string()
+                            } else {
+                                traces
+                                    .into_iter()
+                                    .filter_map(|trace| match trace {
+                                        Trace::Log(s) => Some(s),
+                                        Trace::Label(_) => None,
+                                    })
+                                    .collect::<Vec<_>>()
+                                    .join("\n              ")
+                            }
+                        );
+                    });
+                } else {
+                    // this should allow N scripts to be
+                    let summary: Vec<serde_json::Value> = redeemers
+                        .iter()
+                        .map(|(redeemer, eval)| {
+                            json!({
+                                "mem": redeemer.ex_units.mem as i64,
+                                "cpu": redeemer.ex_units.steps as i64,
+                                "traces": eval.traces().into_iter().filter_map(|trace| match trace {
+                                    Trace::Log(s) => Some(s),
+                                    Trace::Label(_) => None,
+                                }).collect::<Vec<_>>(),
+                            })
+                        })
+                        .collect();
+
+                    println!(
+                        "{}",
+                        serde_json::to_string_pretty(&summary)
+                            .map_err(|_| fmt::Error)
+                            .into_diagnostic()?
+                    );
+                }
             }
             Err(err) => {
                 eprintln!(
-                    "{} {}",
-                    "        Error"
+                    "{:>13} {}",
+                    "Error"
                         .if_supports_color(Stderr, |s| s.red())
                         .if_supports_color(Stderr, |s| s.bold()),
                     err.red()
@@ -167,4 +309,29 @@ pub fn exec(
     }
 
     Ok(())
+}
+
+fn get_override_part<'a>(
+    parts: &'_ mut std::str::Split<'a, &'a str>,
+    when_missing: ScriptOverrideArgumentError,
+) -> Result<&'a str, ScriptOverrideArgumentError> {
+    parts
+        .next()
+        .and_then(|s| if s.is_empty() { None } else { Some(s) })
+        .ok_or(when_missing)
+}
+
+fn decode_script_hash(
+    hex_bytes: &str,
+    when_hex_invalid: impl FnOnce(FromHexError) -> ScriptOverrideArgumentError,
+    when_size_invalid: impl FnOnce(usize) -> ScriptOverrideArgumentError,
+) -> Result<ScriptHash, ScriptOverrideArgumentError> {
+    let bytes = hex::decode(hex_bytes).map_err(when_hex_invalid)?;
+    let size = bytes.len();
+
+    if size != 28 {
+        return Err(when_size_invalid(size));
+    }
+
+    Ok(ScriptHash::from(bytes.as_slice()))
 }